test: Integration Tests

+++ bank-backend/src/main/java/com/gusta/bank/domain/dto/TransactionDTO.java
@@ -5,13 +5,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class TransactionDTO {
     private String to;
-    private BigDecimal value;
+    private double value;
 }
